+++ docs/changelog.rst
@@ -1,6 +1,18 @@
 Changelog
 ==========
 
+v3.0.0
+------
+* Fix that the passed ``port`` keyword argument to ``(Async)YouTubeNotifier.run()`` was ignored when ngrok tunnel was used.
+* Deprecate ``app`` keyword argument for ``(Async)YouTubeNotifier.run()``. It will be removed in version 4.0.0. Pass the ``FastAPI`` instance to the constructor instead.
+* Send an unsubscribe request to YouTube when the notification was for the channel that is not subscribed anymore.
+* Stopping the notifier no longer sends unsubscribe requests.
+* Remove deprecated methods and decorators:
+  * ``(Async)YouTubeNotifier.listener()``
+  * ``(Async)YouTubeNotifier.add_listener()``
+  * ``AsyncYouTubeNotifier.serve()``
+
+
 v2.1.4
 ------